Implement While loop
TODO:
* Fix tests with doubles systematically failing
* Determine offset needed when there are function calls (fibonacci.c)
* Fix assignments not taking expressions for initializations

+++ compiler_tests/custom/custom_float/smaller_and_greater.c
@@ -1,8 +1,8 @@
-int f(double a, double b)
+int f(float a, float b)
 {
     if (a < b)
     {
-        return a <= 4.0l;
+        return a <= 4.0;
     }
-    return b >= 4.0l;
+    return b >= 4.0;
 }
