+++ compiler_tests/custom/custom_float/smaller_and_greater_driver.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int f(double a, double b);
+int f(float a, float b);
 
 int main()
 {
